Add toolchain file to compile host side with clang
This will:
1. Add function that downloads LLVM+Clang 12.0.0 into
'build/common_dependencies' directory, to not download it for each build
configuration, but once and for all build configurations (for host,
device build; for release and debug builds).
2. Add toolchain file which sets the compiler paths to the downloaded
toolchain and adds initial compile flags for C/C++ to emit more errors.

Note that the `CMAKE_TOOLCHAIN_FILE` must be set before the `project`
command. CMake evaluates the variable during `project` command.

We can download the LLVM into a common directory and unpack it there,
because it is a tool run on host, and the target architecture doesn't
affect it.