Add test to static analyse the code with scan-build
This will add a new file with utilities, that runs the macro
EnableClangStaticAnalysis during `include` CMake call.
The macro will add a new test that recompiles the entire project under
scan-build, that is supplied with the LLVM Toolchain downloaded for that
project.
See the comments in the code for more explanation.