Fix scan-build static analysis
The previous approach didn't work properly, because the CC and CXX
supplied by scan-build were not used. After changing the toolchain file
and redesigning the 'clang_static_analysis' target to run cmake's
generate and configure step with scan-build too, it will work.
Previously the generate and configure step were not run under
scan-build. This meant that CC and CXX were not injected to CMake, but
to Make instead (only to the build step). In turn, CC and CXX were not
replaced at all.