Add static analysis with clang-tidy
After installing and using CLion I got excited with built-in clang-tidy.
I have used the same checks CLion does with clang-tidy and added a test
that runs the checks over the code with clang-tidy. This is why I have
also enabled turning Warnings to Errors for clang-tidy - the ctest
command fails with the test 'clang_tidy_static_analysis', when warnings
or errors are found.