Add Loggable concept
Since we can't use dynamic polymorphizm to define Logger interface, I
decided to go for static interfaces. To do that the Loggable concept has
been defined, which expects a log(LogLevel{}) method and overloads for
strings (char arrays, string literals, std::string) and common basic
types. All of the overloads must be defined. They may be defined through
specialization of each function or by a template method.

I have added static assertion tests to check whether the concept is well
defined. The tests will compile a file with the concept used and check
for the compilation result. There are positive and negative tests. It
means that for the former, compilation must pass and for the latter, it
must fail.

I don't understand (yet) how does this concept work. I have used the
solution from the StackOverflow post I have created:
https://stackoverflow.com/questions/69585297/avoiding-implicit-conversion-with-concepts/69587497