Simplify Loggable
I was too enthusiastic about Concepts and overengineered Loggable
concept. From the interface point of view, the MeasurementScheduler
needs it to use it like that:

    logger(LogLevel::info) << "some number: " << 2;

And that's it. No more magic, disabling implicit conversion and return
value specification. Those are details, and the implementer class is
responsible for the details.