Add function which provides ARM GNU Toolchain
It downloads the latest ARM GNU Toolchain for Linux. We will use it
along with the LLVM Toolchain. The LLVM toolchain does not have
precompiled libraries (libc, libm, ...) so we have to provide it on our
own. Those precompiled libraries, for ARM Cortex M4, can be found in the
ARM GNU Toolchain.