Allow to create device executables
This will add a CMake function which is able to create executables that
are strictly targeted to our device. That includes:

* using the linker script generated by CubeMX
* using the startup file generated by CubeMX
* creating "-flash" target that can be used to flash firmware

Miscellanous:
* generation of the HEX file
* printing binary size after build is done

To compile the firmware correctly I had to define the _init() function.
Otherwise lld will fail linking this function, because it is never
defined. I guess that GCC generates it on its own. I have checked the
_init() function generated by GCC and it looks like that:

    push {r3, r4, r5, r6, r7, lr}
    nop
    pop	{r3, r4, r5, r6, r7}
    pop	{r3}
    mov	lr, r3
    bx	lr

Basically it does nothing: calls to push and pop balance each other,
effectively doing no action.