Try compile Catch2 for device
The compilation is successful, but unusable, because the binary size is
too large. I had to hack the FLASH size in the linker script to make the
compilation work and figure out what makes the size explode so much. It
turns out to be the Catch2 CLI parser which takes much space.

I debugged the binary size with bloaty:
https://github.com/google/bloaty. I had to compile the firmware with
the debug symbols enabled:

    cmake ../../device -DCMAKE_BUILD_TYPE=Debug

Compiling Catch2 without cout, cin and cerr included didn't help much.
Also disabling the colourful output didn't decrease the binary size a
lot, which was in the end:

    Section sizes of the executable binary
       text    data     bss     dec     hex filename
     424408    2316   12268  438992   6b2d0 /home/kacper/Workspace/Aura/build/build_device_release/tests/device/dummy_test