Log simple log over SerialLogger
This is successful:
1. I connected PA2 to RX pin of USB-UART converter.
2. I attached a terminal like this:

    screen /dev/ttyUSB1 115200

3. Flashed the firmware with:

     cmake --build . --target test_serial_logger-flash

4. Could see the line:

    INFO: some simple log

in the screen session.

---

This implements SerialLogger which initializes the peripheral on
creation and deinitializes it on destruction. HAL API is used to talk to
the peripheral.

SerialLoggerProxy is returned by the log() function to:

* forbid usage of operator<< on SerialLogger directly.
* enable RAII for calls to log():

    SerialLogger{}.log(LogLevel::info) << "some log" << 3232 << 'c';

When log() creates the SerialLoggerProxy then we can prepare the line
(e.g. print "INFO: ") from constructor of SerialLoggerProxy. The object
will be propagated as an lvalue to each operator<<() call. When reaching
";" it will be destructed and the log line can be tear down (e.g. "\r\n"
can be outputted).