Fix app crash
Because of this missing flag the "data" section was stripped and there
were errors during HAL Initialization.

I figured it out by inspecting the calls to HAL_Init(). Within it there
is call to HAL_InitTick, where "SystemCoreClock" variable is used. It
had invalid value, which was random during various calls.

Then I used llvm-objdump, bloaty and map file to check when it is
initialized and with what value. This is done in startup_stm32*.s file,
inside CopyDataInit. It's the place where global variables with
non-zero (non-default) initial value are initialized. The place where
the initial value should lay (inside the .data section in FLASH) was
stripped (or cut).

I experimented with multiple compile options and '-flto' turned out to
be the one which was problematic. It seems that this flag must be
enabled both, for compilation and linkage stage. I don't know why is
that.