Add initial script capturing serial in background
This script was inteded to serve serial over TCP, but in the background,
as a daemon. This is wrong approach and I abandon it, because I thought
that it will be possible for CMake to read TCP in a non-blocking manner
and use then PASS_REGULAR_EXPRESSION test property. This is impossible,
so this script will be reverted.

We will use a python3 script that performs the regex matching on its
own.