Add first serial_logger test
This simple_log test will check whether a simple info log is printed. A
new CMake function is added to simplify adding more of the tests of the
serial logger. Thanks to it we will test each functionality with a
single binary and will match against expected output with PASS_REGEX.

The test_serial_logger.cpp will be split to multiple executables, to
test each functionality with a single binary, thus with a single simple
matching REGEX.

CreateSerialLoggerTestExpectingOutput has its documentation defined in
details.