Fix printing large floats
I misunderstood the option MAX_INTEGRAL_DIGITS_FOR_DECIMAL, because I
thought that it is the max number of digits in the decimal part (on the
left-side of the '.' in the number).

When the float is large and the integral part has more than 9 digits,
then exponential format (scientific) is used. Since we disabled that
option during compilation, then no float will be printed. That's wrong.
It increases the binary size by 2kB but, that's fine.