Add toolchain file for ARM GNU GCC toolchain
To build with this toolchain run:

    mkdir -p build/build_device_arm_gnu_release
    cd build/build_device_arm_gnu_release
    cmake \
        -DCMAKE_TOOLCHAIN_FILE=../../cmake/arm_gnu_gcc_device_build_toolchain_file.cmake \
        ../../device
    cmake --build .

For debug build:

    mkdir -p build/build_device_arm_gnu_debug
    cd build/build_device_arm_gnu_debug
    cmake \
        -DCMAKE_TOOLCHAIN_FILE=../../cmake/arm_gnu_gcc_device_build_toolchain_file.cmake \
        -DCMAKE_BUILD_TYPE=Debug
        ../../device
    cmake --build .

The layout of the file is scrapped from the file for Clang. The file is
less complicated, because the toolchain knows which libraries to pull
and which paths to use.

There is no libc-nano used, because then the exceptions will not work.