Decrease binary size by linking custom terminate
This overrides custom terminate function, the function which is called
when uncaught exception is handled. Standard C++ library's default
terminate handler takes 50 kB. That's way too much.

I had to add it with target_sources, because linking it with
target_link_libraries, by adding the custom_terminate.cpp to the
'device_specific' target, lead to not linking the custom
__terminate_handler to the final executable.