Fix linker flags when linking GCC libs with clang
Exceptions didn't work when clang was used to link libraries built with
ARM GNU Toolchain. See: https://lists.llvm.org/pipermail/llvm-dev/2021-November/153810.html

I have fixed few flags to be properly propagated to the linker by
clang++. Other fix is that we use the "crt*" objects which are normally
used by GCC. GCC is figuring it out on its own, basing on the target
triplet. The most important fix is use of the flag:

    " -Wl,--target2=rel"

This makes the linker not generate "got" section and all the jumps are
resolved correctly.