Work-around for clang build: multiple symbol def
This is a temporary workaround. When using clang the symbol
__cxxabiv1::__terminate_handler will be defined twice: once in the
libstdc++ and second time in the custom_terminate.cpp. We would like to
use the one in custom_terminate.cpp, but I don't have solution for it:

See: https://stackoverflow.com/questions/70153260/turning-symbol-weak-at-link-time-to-resolve-multiple-definitions

This is a post where I ask for the solution.