Drop -nodefaultlibs, -nostdlib and -sysroot
We drop it for two reasons:
1. We want the linker driver to resolve paths to standard libraries on
its own. This is due to overrides of standard library functions. If we
use -nostdlib, then we have to specify '-lgcc -lstdc++' on our own. This
means that we control the linking order, thus the custom objects may be
linked after standard libraries. It leads to multiple definition error.
The custom objects, with custom standard library overrides, must be
linked before the standard libraries.
2. Using sysroot without -nostdlib, leads to linking generic standard
libraries from the 'lib' subdirectory from the ARM GNU Toolchain,
instead of pulling the standard libraries cross-compiled for our target
architecture (v7e-m+fp/hard/ etc.).

This, in turn, leads to:
1. Include flags must be scrapped from the arm-none-eabi-gcc, since we
don't use 'sysroot'.
2. We must ask clang to decide on his own to pull proper standard
librarues with -rtlib and -stdlib flags.

Now, we will test whether the custom standard library symbols are pulled
correctly.