Use FreeRTOS-Kernel main CMakeLists.txt
The main CMakeLists.txt within the FreeRTOS-kernel is well written CMake
script, which provides the FreeRTOS library with all the bits and pieces
and is configurable via CMake CACHE variables. Basically, it does the
same thing like the function from CMake helpers. Therefore, the CMake
helpers are no longer needed.