Fix scan-build static analysis
Either new version of CMake broke 'scan-build' find_program(), or my
local installation was wrong, because the static_code_analysis target
test failed during compilation. I had to fix the compiler paths and the
target to run the test properly.
All the external dependencies are excluded from the scan.
Ideally, we would like the scan-build to run fast, only on the project
files we want to skip the CMake generate and configure step, to skip
linking and to not build the files we don't want to scan.